feat: add copy-to-clipboard for Bash commands via reusable CopyableContent
- Extract copy-to-clipboard logic from MessageBlock into reusable CopyableContent component
- Refactor MessageBlock to use CopyableContent wrapper
- Add copy functionality to Bash tool commands in AgentChain
- Support customizable button positioning and tooltip text
