+++ apps/agor-ui/src/components/CopyableContent/index.ts
@@ -1,0 +1,2 @@
+export type { CopyableContentProps } from './CopyableContent';
+export { CopyableContent } from './CopyableContent';
